♻️ refactor: プレビューパネル描画ロジックを独立モジュールへ分離
`renderPreviewPanel()`を`preview.js`から新規`preview_panel.js`へ抽出し、`events.js`の`renderApp()`でプレビューパネル表示時に自動更新するよう修正
